+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Message::into_packet` function.
 - `Message::code_index_number` function.
 - `CodeIndexNumber::event_size()` function.
+- `CableNumber::Cable0` as default value.
 - Re-exports of common items in crate root.
 
 ### Changed